[B] Fix automated navigation issues
These ∆s fix the issues with the nav:

- Generate the nav for all visible non-site-root pages
- Clean up to use native model methods for the most part
- Fix a few wee template issues

There's a fix for Contentment that will allow us to nest the pages and
get the URIs we want. We can integrate that with Contentment, but it
warrants deciding if we want to require all pages to be children of
site_root (if there is one). Currently, Contentment is designed so that
site_root just means that a particular page has no slug, not that it's a
parent of all the other pages.

[#88]