[B] Bump contentment with new URI handling
This is an interim change that will allow pages to be stored in the
rootline for the site. It also will require that they are, for the
routing to work. A more robust fix is in the works for Contentment, and
we should integrate that when ready. For this reason, I've used the
branch specifier in composer.json for now, to make it clear this is an
interim solution.

[Fixes #88]